refactor(engine): ReleaseProcessor 10a — preview/outcome + streaming/press calc cluster
Phase 2 engine-seams PR-10, commit A (of two on this branch). Extracts the
release-domain PREVIEW/OUTCOME leaf cluster from GameEngine into a new
stateless ReleaseProcessor (shared/engine/processors/ReleaseProcessor.ts),
following the established WeekContext processor pattern (PR-5..9).

Moved VERBATIM (this. rebound to ctx only):
  calculateReleasePreview (largest single method, ~250 ln)
  calculateSophisticatedReleaseOutcome
  calculatePressOutcome, calculateStreamingOutcome, calculateOngoingSongRevenue
  + preview-private helpers calculateReleaseRisks, getChannelSynergiesFromBalance,
    generateReleaseRecommendations

Preview-first ordering: the weekly pipeline (moved in commit B) depends on
calculateSophisticatedReleaseOutcome -> calculateReleasePreview, so the preview
cluster is the leaf and must land first for each commit to be self-consistent
and independently byte-identical.

Same-signature engine delegates kept for external/public callers:
  calculateReleasePreview — server/routes/releases.ts:429 preview endpoint
    (guarded by tests/endpoints/releases.characterization.test.ts)
  calculateStreamingOutcome / calculatePressOutcome / calculateOngoingSongRevenue

RNG: the press-pickup roll draws from the engine's single seeded stream via
ctx.getRandom; draw order unchanged. Awareness breakthrough uses a deterministic
Math.sin hash (not the RNG stream).

Gates: npm run check clean; golden master 8/8 zero-update (x2);
releases.characterization 15/15 green.
